feat(interrupts): add interrupts to io.c and test functions in test.c
STM32 uses a similar process to other mcu's in the interrupt process.
Instead of using ports to specify pins, it uses the pins directly. This
means that there are only 16 possible pins that can be configured as an
interrupt with 7 handlers in total. The process of configuring an
interrupt starts by enabling the clock. Then, a pin is sslected and
saved in teh SYSCFG register. Afterwards, the interrupt is disabled and
a trigger is saved. Then, the interrupt is enabled.

Interrupts must be globally enabled and are done so in the mcu_init
function.

The process of using them as a user in the test.c function is simple:
1. io_configure_interrupt
2. io_enable_interrupt